Add requireConfig() override for CLI
added a defined check to see if another php script already injected the database information to the putenv() method to skip over requiring a php configuration file. For example, this will allow the vfmagento cli script to pull the XML configuration from the magento installation by default, inject it by using putenv()  without having to specify the credentials for the core library.